+++ src/Oidc/OidcTokenSet.php
@@ -6,10 +6,11 @@ final readonly class OidcTokenSet
 {
     public function __construct(
         public string $accessToken,
-        public string $idToken,
+        public ?string $idToken,
         public ?string $refreshToken,
         public int $expiresIn,
         public string $tokenType,
+        public ?string $scope = null,
     ) {
     }
 }